Add a help markdown file for the History Switcher
This commit only adds a `selector_modal_switch.md` which explains the usage of the _Switch to History_ modal, from `SelectorModal.vue`.
There is another version of `SelectorModal.vue` that takes the boolean prop `multiple`, which changes it into the History Multiview selector (accessible through the button on the top right of the multiview), and we will need to add another `.md` help file for that (named `selector_modal_multiview.md`).
In Galaxy, we will then conditionally fetch the appropriate help `.md` based on the component's prop.